sesiones y configuracion apache

+++ artistas/eliminar_artistas.php
@@ -1,4 +1,7 @@
 <?php
+//Verificar si existe una sesion:
+session_start();
+if(isset($_SESSION['valida']) && $_SESSION['valida'] == true){
 	include 'conexion.php';
 	$id= $_POST['id'];
 	$query = "delete from artistas where artista_id='$id';";
@@ -8,6 +11,11 @@
 	}else{
 		echo "No se elimino";
 	}	
+
+}
+else{
+	header('Location: ../index.php?error=2');
+}
 
 
 ?>
